requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-jose>=3.0.0
passlib>=1.7.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
jinja2>=3.0.0
aiofiles>=0.7.0
sqlalchemy>=1.4.0